updt: inline and minify html

+++ src/takeScreenshot.cjs
@@ -2,9 +2,15 @@ const puppeteer = require("puppeteer");
 const fs = require("fs");
 const crypto = require("crypto");
 
+const sleep = (ms) => new Promise((resolve) => setTimeout(resolve, ms));
+
 const SOURCE_PATH = "src/";
 const RELATIVE_RESOURCE_PATH = "dist/";
 const RESOURCE_PATH = SOURCE_PATH + RELATIVE_RESOURCE_PATH;
+let NOCACHE = false;
+if (process.argv.length > 2) {
+	NOCACHE = process.argv[2] == "nocache";
+}
 
 function createHashFromURL(url) {
 	const hash = crypto.createHash("sha256");
@@ -14,12 +20,14 @@ function createHashFromURL(url) {
 }
 
 (async () => {
-	fs.rmSync(RESOURCE_PATH, { recursive: true, force: true });
-	fs.mkdirSync(RESOURCE_PATH);
-	fs.writeFileSync(
-		SOURCE_PATH + "resources.js",
-		"export const IMAGEMAP = {};"
-	);
+	if (!NOCACHE) {
+		fs.rmSync(RESOURCE_PATH, { recursive: true, force: true });
+		fs.mkdirSync(RESOURCE_PATH);
+		fs.writeFileSync(
+			SOURCE_PATH + "resources.js",
+			"export const IMAGEMAP = {};"
+		);
+	}
 
 	const browser = await puppeteer.launch({
 		args: ["--no-sandbox", "--disable-setuid-sandbox"],
@@ -28,33 +36,35 @@ function createHashFromURL(url) {
 
 	var imageMap = {};
 
-	// Event listener for response to capture resources
-	page.on("response", async (response) => {
-		const headers = response.headers();
-		const contentType = headers["content-type"];
-		if (contentType && contentType.startsWith("image")) {
-			const addr = response.remoteAddress();
-			if (addr.ip == "127.0.0.1" && addr.port == 8000) {
-				return;
+	if (!NOCACHE) {
+		// Event listener for response to capture resources
+		page.on("response", async (response) => {
+			const headers = response.headers();
+			const contentType = headers["content-type"];
+			if (contentType && contentType.startsWith("image")) {
+				const addr = response.remoteAddress();
+				if (addr.ip == "127.0.0.1" && addr.port == 8000) {
+					return;
+				}
+				const url = response.url();
+				const content = await response.buffer();
+				const extension = contentType.split("/")[1];
+
+				// Save the image
+				const fileName = createHashFromURL(url) + "." + extension;
+				fs.writeFileSync(RESOURCE_PATH + fileName, content);
+				imageMap[url] = {
+					file: RELATIVE_RESOURCE_PATH + fileName,
+					lastModified: headers["last-modified"] || "",
+					etag: headers["etag"] || "",
+				};
+
+				console.log(
+					`Image saved: ${fileName}: ${url} ${headers["last-modified"]} ${headers["etag"]}`
+				);
 			}
-			const url = response.url();
-			const content = await response.buffer();
-			const extension = contentType.split("/")[1];
-
-			// Save the image
-			const fileName = createHashFromURL(url) + "." + extension;
-			fs.writeFileSync(RESOURCE_PATH + fileName, content);
-			imageMap[url] = {
-				file: RELATIVE_RESOURCE_PATH + fileName,
-				lastModified: headers["last-modified"] || "",
-				etag: headers["etag"] || "",
-			};
-
-			console.log(
-				`Image saved: ${fileName}: ${url} ${headers["last-modified"]} ${headers["etag"]}`
-			);
-		}
-	});
+		});
+	}
 
 	// Replace with the actual URL of your Tailwind page
 	const url = "http://localhost:8000";
@@ -62,6 +72,7 @@ function createHashFromURL(url) {
 
 	await page.setViewport({ width: 2100, height: 1440 });
 	await page.waitForNetworkIdle();
+	await sleep(1000);
 
 	// Get the page height
 	const pageHeight = await page.evaluate(() => document.body.scrollHeight);
@@ -99,35 +110,39 @@ function createHashFromURL(url) {
 		screenshots.push(screenshot);
 	}
 
-	// Find out the element corresponding to each image in the imageMap
-	for (const [url, data] of Object.entries(imageMap)) {
-		const element = await page.evaluateHandle((url) => {
-			return document.querySelector(`img[src="${url}"]`);
-		}, url);
+	if (!NOCACHE) {
+		// Find out the element corresponding to each image in the imageMap
+		for (const [url, data] of Object.entries(imageMap)) {
+			const element = await page.evaluateHandle((url) => {
+				return document.querySelector(`img[src="${url}"]`);
+			}, url);
 
-		if (element) {
-			const boundingBox = await element.boundingBox();
-			console.log(
-				`Element found for ${url} in bounding box:`,
-				boundingBox
-			);
-			if (boundingBox) {
-				const height = Math.ceil(boundingBox.height);
-				const width = Math.ceil(boundingBox.width);
+			if (element) {
+				const boundingBox = await element.boundingBox();
+				console.log(
+					`Element found for ${url} in bounding box:`,
+					boundingBox
+				);
+				if (boundingBox) {
+					const height = Math.ceil(boundingBox.height);
+					const width = Math.ceil(boundingBox.width);
 
-				// Resize the original image to fit the bounding box
-				const fileName = data.file;
-				const convertedFile = fileName.split(".")[0] + ".avif";
-				const originalImage = fs.readFileSync(SOURCE_PATH + fileName);
+					// Resize the original image to fit the bounding box
+					const fileName = data.file;
+					const convertedFile = fileName.split(".")[0] + ".avif";
+					const originalImage = fs.readFileSync(
+						SOURCE_PATH + fileName
+					);
 
-				const sharp = require("sharp");
-				await sharp(originalImage)
-					.resize(width, height)
-					.toFile(SOURCE_PATH + convertedFile);
+					const sharp = require("sharp");
+					await sharp(originalImage)
+						.resize(width, height)
+						.toFile(SOURCE_PATH + convertedFile);
 
-				data.file = convertedFile;
-				console.log("Saved resized image:", data.file);
-				fs.rmSync(SOURCE_PATH + fileName, { force: true });
+					data.file = convertedFile;
+					console.log("Saved resized image:", data.file);
+					fs.rmSync(SOURCE_PATH + fileName, { force: true });
+				}
 			}
 		}
 	}
@@ -139,9 +154,12 @@ function createHashFromURL(url) {
 	screenshots.forEach((screenshot, index) => {
 		fs.writeFileSync(`part${index}.png`, screenshot);
 	});
-	console.log(JSON.stringify(imageMap));
-	fs.writeFileSync(
-		SOURCE_PATH + "resources.js",
-		"export const IMAGEMAP = " + JSON.stringify(imageMap, null, 4) + ";"
-	);
+
+	if (!NOCACHE) {
+		console.log(JSON.stringify(imageMap));
+		fs.writeFileSync(
+			SOURCE_PATH + "resources.js",
+			"export const IMAGEMAP = " + JSON.stringify(imageMap, null, 4) + ";"
+		);
+	}
 })();
